# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinanIes",
        "contact": {
            "email": "derex@outlook.com.br"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/v1/permission": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "get all permissions",
                "description": "get all permissions",
                "operationId": "550d293092d8f45fe8379222bb165be6",
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Permission"
                ],
                "summary": "register a new permission",
                "description": "register a new permission",
                "operationId": "7ba7bb2d9c5e353725c90f6ca637f156",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "description",
                                        "type": "string"
                                    },
                                    "view": {
                                        "description": "view",
                                        "type": "boolean"
                                    },
                                    "edit": {
                                        "description": "edit",
                                        "type": "boolean"
                                    },
                                    "delete": {
                                        "description": "delete",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/permission/{permission}": {
            "put": {
                "tags": [
                    "Permission"
                ],
                "summary": "update a permission by id",
                "description": "update a permission by id",
                "operationId": "updatePermission",
                "parameters": [
                    {
                        "name": "permission",
                        "in": "path",
                        "description": "permission identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "description",
                                        "type": "string"
                                    },
                                    "view": {
                                        "description": "view",
                                        "type": "boolean"
                                    },
                                    "edit": {
                                        "description": "edit",
                                        "type": "boolean"
                                    },
                                    "delete": {
                                        "description": "delete",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Permission"
                ],
                "summary": "delete a permission by id",
                "description": "delete a permission by id",
                "operationId": "deletePermission",
                "parameters": [
                    {
                        "name": "permission",
                        "in": "path",
                        "description": "permission identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/permission/{permission}/attach": {
            "post": {
                "tags": [
                    "Permission"
                ],
                "summary": "associate permission to category",
                "description": "associate permission to category",
                "operationId": "3344e44e418f061a353220e404f9de1a",
                "parameters": [
                    {
                        "name": "permission",
                        "in": "path",
                        "description": "permission identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "description": "category id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "Bearer"
            }
        }
    }
}